requirements.txt
azure-functions
tensorflow
tensorflow-addons 
numpy
Pillow
opencv-python
requests